Remove duplicate instance of regex
